% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.methods.R
\name{bread.gmnl}
\alias{bread.gmnl}
\title{Bread for Sandwiches}
\usage{
\method{bread}{gmnl}(x, ...)
}
\arguments{
\item{x}{a fitted model of class \code{gmnl}.}

\item{...}{other arguments when \code{bread} is applied to another
class object.}
}
\value{
The covariance matrix times observations
}
\description{
Computes the ``bread'' of the sandwich covariance matrix for objects of class \code{gmnl}.
}
\details{
For more information see \code{\link[sandwich]{bread}} from the package \pkg{sandwich}.
}
\references{
Zeileis A (2006), Object-oriented Computation of Sandwich 
Estimators. Journal of Statistical Software, 16(9), 1--16.
}
