% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.methods.R
\name{estfun.gmnl}
\alias{estfun.gmnl}
\title{Gradient for Observations}
\usage{
\method{estfun}{gmnl}(x, ...)
}
\arguments{
\item{x}{a fitted model of class \code{gmnl}.}

\item{...}{other arguments. Ignored.}
}
\value{
The gradient matrix of dimension \eqn{n \times K}
}
\description{
It extracts the gradient for each observation evaluated at the estimated parameters for an object of class \code{gmnl}.
}
\details{
For more information see \code{\link[sandwich]{estfun}} from package \pkg{sandwich}.
}
\references{
Zeileis A (2006), Object-oriented Computation of Sandwich 
Estimators. Journal of Statistical Software, 16(9), 1--16.
}
