% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmnl.methods.R
\name{getSummary.gmnl}
\alias{getSummary.gmnl}
\title{Get Model Summaries for Use with "mtable"}
\usage{
getSummary.gmnl(obj, alpha = 0.05, ...)
}
\arguments{
\item{obj}{a \code{gmnl} object,}

\item{alpha}{level of the confidence intervals,}

\item{...}{further arguments,}
}
\description{
A generic function to collect coefficients and summary statistics from a \code{gmnl} object. It is used in \code{mtable}.
}
\details{
For more details see package \pkg{memisc}
}
